+++ app/src/main/java/com/example/bangwool/ui/mypage/MyPageFragment.kt
@@ -1,14 +1,13 @@
 package com.example.bangwool.ui.mypage
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.bangwool.AboutDialogUtils
-import com.example.bangwool.LoginActivity
-import com.example.bangwool.WithdrawDialogUtils
 import com.example.bangwool.databinding.FragmentMypageBinding
+import com.example.bangwool.ui.login.LoginActivity
 
 class MyPageFragment : Fragment() {
     lateinit var binding: FragmentMypageBinding
