+++ app/src/main/res/drawable/dialog_about_background.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
-    <solid android:color="@color/white"
-        />
-    <corners android:radius="16dp"
-        />
+    <solid android:color="@color/white" />
+    <corners android:radius="16dp" />
 </shape>
